Correcciones. y marge de FormRecetas + Mis Recetas
